requirements.txt
aiogram==3.5.0
SQLAlchemy==2.0.27
asyncpg==0.29.0
pydantic==2.6.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
httpx==0.27.2
